+++ src/ImageDimensions.php
@@ -13,6 +13,7 @@ declare(strict_types=1);
 namespace Contao\Image;
 
 use Contao\ImagineSvg\RelativeBoxInterface;
+use Contao\ImagineSvg\SvgBox;
 use Contao\ImagineSvg\UndefinedBoxInterface;
 use Imagine\Image\BoxInterface;
 
@@ -59,11 +60,13 @@ class ImageDimensions
         }
 
         if (null === $relative) {
-            $relative = $size instanceof RelativeBoxInterface;
+            $relative = $size instanceof RelativeBoxInterface
+                || ($size instanceof SvgBox && SvgBox::TYPE_ASPECT_RATIO === $size->getType());
         }
 
         if (null === $undefined) {
-            $undefined = $size instanceof UndefinedBoxInterface;
+            $undefined = $size instanceof UndefinedBoxInterface
+                || ($size instanceof SvgBox && SvgBox::TYPE_NONE === $size->getType());
         }
 
         $this->size = $size;