+++ tests/ImageDimensionsTest.php
@@ -14,6 +14,7 @@ namespace Contao\Image\Tests;
 
 use Contao\Image\ImageDimensions;
 use Contao\ImagineSvg\RelativeBoxInterface;
+use Contao\ImagineSvg\SvgBox;
 use Contao\ImagineSvg\UndefinedBoxInterface;
 use Imagine\Image\BoxInterface;
 use PHPUnit\Framework\TestCase;
@@ -51,7 +52,11 @@ class ImageDimensionsTest extends TestCase
 
         $this->assertTrue($dimensions->isRelative());
 
-        $size = $this->createMock(RelativeBoxInterface::class);
+        if (class_exists(SvgBox::class)) {
+            $size = SvgBox::createTypeAspectRatio(100, 100);
+        } else {
+            $size = $this->createMock(RelativeBoxInterface::class);
+        }
         $dimensions = new ImageDimensions($size);
 
         $this->assertTrue($dimensions->isRelative());
@@ -72,7 +77,11 @@ class ImageDimensionsTest extends TestCase
 
         $this->assertTrue($dimensions->isUndefined());
 
-        $size = $this->createMock(UndefinedBoxInterface::class);
+        if (class_exists(SvgBox::class)) {
+            $size = SvgBox::createTypeNone();
+        } else {
+            $size = $this->createMock(UndefinedBoxInterface::class);
+        }
         $dimensions = new ImageDimensions($size);
 
         $this->assertTrue($dimensions->isUndefined());