+++ tests/ResizerTest.php
@@ -20,6 +20,7 @@ use Contao\Image\ResizeCoordinates;
 use Contao\Image\ResizeOptions;
 use Contao\Image\Resizer;
 use Contao\ImagineSvg\Imagine as SvgImagine;
+use Contao\ImagineSvg\SvgBox;
 use Contao\ImagineSvg\UndefinedBox;
 use Imagine\Gd\Imagine as GdImagine;
 use Imagine\Image\Box;
@@ -349,7 +350,9 @@ class ResizerTest extends TestCase
         $image = $this->createMock(Image::class);
         $image
             ->method('getDimensions')
-            ->willReturn(new ImageDimensions(new UndefinedBox()))
+            ->willReturn(new ImageDimensions(
+                class_exists(SvgBox::class) ? SvgBox::createTypeNone() : new UndefinedBox()
+            ))
         ;
 
         $image